NavBar related updates
- Include the NavBar on every page
- Normalize the headers on each page
- Normalize the navigation buttons on each page
- Update styling and div layout to simplify things and work with new layouts
- Update border of heading and naviagtion buttons
- Update positioning of alert when there is an error loading a config file
