+++ react-app/src/components/Extract.js
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { Button, Col, Form, Row } from 'react-bootstrap';
 
-import Navbar from './Navbar';
 import FilePicker from './FilePicker';
 import LinkButton from './LinkButton';
 
@@ -86,11 +85,10 @@ function Extract(props) {
     setSubmitted(!submitted);
   }
   return (
-    <div className="page-container">
-      <Navbar />
+    <>
       <h1 className="page-title">Extract New</h1>
       {!submitted && (
-        <div className="page-container">
+        <div className="page-content">
           <Form className="form-container">
             <Row>
               <Col>
@@ -172,7 +170,7 @@ function Extract(props) {
           </div>
         </div>
       )}
-    </div>
+    </>
   );
 }
