+++ react-app/src/components/ExtractionError.js
@@ -1,12 +1,10 @@
 import React from 'react';
-import Navbar from './Navbar';
 import LogList from './LogList';
 import LinkButton from './LinkButton';
 
 function ExtractionError(props) {
   return (
     <div>
-      <Navbar />
       <h1 className="page-title">An error occurred during extraction.</h1>
       <p className="page-text text-centered">
         For more information, view the logs below and/or run extraction again with "Log output debugging information"
