+++ react-app/src/components/Home.js
@@ -3,13 +3,11 @@ This is the landing page that will display when users first open the app. It has
 */
 
 import React from 'react';
-import Navbar from './Navbar';
 import LinkButton from './LinkButton';
 
 function Home() {
   return (
     <div>
-      <Navbar />
       <h1 className="page-title">Extractor UI</h1>
       <div className="button-container">
         <LinkButton className="vertical-menu-button" variant="outline-secondary" text="Extract New" path="/extract" />
