+++ react-app/src/components/OutputDisplay.js
@@ -5,7 +5,7 @@ import SaveOutputForm from './SaveOutputForm';
 
 function OutputDisplay(props) {
   return (
-    <div>
+    <>
       {props.id === null && !props.showLogs && !props.showSaveForm && (
         <p className="page-text text-centered padding-100">
           Select a patient to view their information or click on "Logger Messages" to view the logger messages.
@@ -13,14 +13,11 @@ function OutputDisplay(props) {
       )}
       {props.showLogs && !props.showSaveForm && (
         <div>
-          <h3 className="page-subtitle sticky-item">Log File</h3>
           <LogList loggedMessages={props.loggedMessages} listType="log-list-result-page" />
         </div>
       )}
       {props.id !== null && !props.showLogs && !props.showSaveForm && (
-        <div className="patient-json-display">
-          <ReactJson src={props.patientJson} collapsed={4} />
-        </div>
+        <ReactJson src={props.patientJson} collapsed={4} />
       )}
       {props.showSaveForm && (
         <SaveOutputForm
@@ -29,7 +26,7 @@ function OutputDisplay(props) {
           loggedMessages={props.loggedMessages}
         />
       )}
-    </div>
+    </>
   );
 }
