+++ react-app/src/components/ResultSidebar.js
@@ -1,21 +1,8 @@
 import React from 'react';
-import { Button, Accordion } from 'react-bootstrap';
-import { useHistory } from 'react-router-dom';
-import ResultHeader from './ResultHeader';
+import { Accordion } from 'react-bootstrap';
 import Result from './Result';
 
 function ResultSidebar(props) {
-  const history = useHistory();
-
-  function onExitResultPage() {
-    // reset data values and return to home page
-    history.push('/extract');
-  }
-
-  function onSave() {
-    props.setShowSaveForm(true);
-  }
-
   function getLoggerStats() {
     let errors = 0;
     let warnings = 0;
@@ -49,33 +36,22 @@ function ResultSidebar(props) {
   ));
 
   return (
-    <div className="sidebar d-flex flex-column">
-      <ResultHeader />
-      <div className="sidebar-interior">
-        <Accordion defaultActiveKey="0" flush>
-          <Accordion.Item eventKey={-1}>
-            <Accordion.Header
-              onClick={() => {
-                props.setPatientID(null);
-                props.setShowLogs(true);
-                props.setShowSaveForm(false);
-              }}
-            >
-              Logger Messages
-            </Accordion.Header>
-            {getLoggerStats()}
-          </Accordion.Item>
-          {list}
-        </Accordion>
-      </div>
-      <div className="sidebar-button-container">
-        <Button className="generic-button" size="lg" variant="outline-secondary" onClick={onExitResultPage}>
-          Exit
-        </Button>
-        <Button className="generic-button" siz="lg" variant="outline-secondary" onClick={onSave}>
-          Save
-        </Button>
-      </div>
+    <div className="sidebar">
+      <Accordion defaultActiveKey="0" flush>
+        <Accordion.Item eventKey={-1}>
+          <Accordion.Header
+            onClick={() => {
+              props.setPatientID(null);
+              props.setShowLogs(true);
+              props.setShowSaveForm(false);
+            }}
+          >
+            Logger Messages
+          </Accordion.Header>
+          {getLoggerStats()}
+        </Accordion.Item>
+        {list}
+      </Accordion>
     </div>
   );
 }
