+++ react-app/src/stylesheets/LogList.scss
@@ -1,6 +1,4 @@
 .log-list-result-page {
-  margin-top: 80px;
-  max-height: 90vh;
   overflow-y: scroll;
 }
