+++ react-app/src/stylesheets/ResultPage.scss
@@ -10,35 +10,10 @@
   border-right-width: 1px;
   box-sizing: border-box;
   color: rgb(33, 37, 41);
-  display: flex;
-  flex-direction: column;
+  overflow: scroll;
   font-size: 16px;
   font-weight: 400;
-  height: 100vh;
-  line-height: 24px;
-  position: sticky;
-  top: 0px;
-  z-index: 2234;
-}
-
-.sidebar-interior {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  height: 100vh;
-  overflow-y: scroll;
-  padding-right: 0px;
-}
-
-.sidebar-button-container {
-  display: flex;
-  justify-content: space-between;
-  margin-left: 20px;
-  margin-right: 20px;
-}
-
-.patient-json-display {
-  margin-top: 30px;
+  height: 100%;
 }
 
 .result-header {
@@ -55,4 +30,9 @@
   overflow-wrap: break-word;
   margin-bottom: 0px;
   max-width: 190px;
+}
+
+.full-height-scroll {
+  overflow: scroll;
+  height: 100%;
 }
