+++ react-app/src/stylesheets/_variables.scss
@@ -1,5 +1,6 @@
 $light-shade: #f1f3f1;
 $light-accent: #9c8b90;
+$lighter-accent: #ece9eb;
 $main-color: #c3906a;
 $dark-accent: #69738d;
 $dark-shade: #40273c;
